[Drastic] reusable attendance table --change
